add test for initial prompt, reorg files into src/test/build directories, and update .gitignore and makefile
