+++ .gitignore
@@ -52,3 +52,5 @@ Mkfile.old
 dkms.conf
 
 pcshell
+
+build/*.txt
