+++ .vscode/settings.json
@@ -1,0 +1,8 @@
+{
+    "files.associations": {
+        "unity.h": "c",
+        "bitset": "c",
+        "cstdio": "c",
+        "shell.h": "c"
+    }
+}
