+++ src/main.c
@@ -1,0 +1,6 @@
+#include "shell.h"
+
+int main(void) {
+    execute_shell();
+    return 0;
+}
