+++ src/shell.c
@@ -1,11 +1,10 @@
-
 #include <stdbool.h>
 #include <stdio.h>
+#include "shell.h"
 
-int main(void) {
+void execute_shell(void) {
     char command[50];
     printf("pcshell> ");
     scanf("%s", command);
     printf("%s\n", command);
-    return 0;
 }
