+++ src/shell.h
@@ -1,0 +1,8 @@
+#ifndef SHELL_H
+#define SHELL_H
+
+#include <stddef.h>
+
+void execute_shell(void);
+
+#endif
